+++ MediaPlayerWithTest.Business/src/Service/MediaService.cs
@@ -1,5 +1,6 @@
 using MediaPlayerWithTest.Domain.src.RepositoryInterface;
 using MediaPlayerWithTest.Business.src.ServiceInterface;
+using MediaPlayerWithTest.Domain.src.Core;
 
 namespace MediaPlayerWithTest.Business.src.Service
 {
@@ -12,24 +13,30 @@ namespace MediaPlayerWithTest.Business.src.Service
             _mediaRepository = mediaRepository;
         }
 
-        public void CreateNewFile(string fileName, string filePath, TimeSpan duration)
+        public MediaFile CreateNewFile(string fileName, string filePath, TimeSpan duration)
         {
-            _mediaRepository.CreateNewFile(fileName, filePath, duration);
+            return _mediaRepository.CreateNewFile(fileName, filePath, duration);
         }
 
-        public void DeleteFileById(int id)
+        public bool DeleteFileById(int id)
         {
+            var foundFile = _mediaRepository.GetFileById(id);
+            if (foundFile == null)
+            {
+                throw new Exception("ID not valid");
+            }
             _mediaRepository.DeleteFileById(id);
+            return true;
         }
 
-        public void GetAllFiles()
+        public IEnumerable<MediaFile> GetAllFiles()
         {
-            _mediaRepository.GetAllFiles();
+            return _mediaRepository.GetAllFiles();
         }
 
-        public void GetFileById(int id)
+        public MediaFile GetFileById(int id)
         {
-            _mediaRepository.GetFileById(id);
+            return _mediaRepository.GetFileById(id);
         }
     }
 }
